+++ robotframework-ls/src/robotframework_debug_adapter/launch_process.py
@@ -725,7 +725,11 @@ class LaunchProcess(object):
     def disconnect(self, disconnect_request: DisconnectRequest) -> None:
         from robocorp_ls_core.options import is_true_in_env
 
-        if is_true_in_env("RFLS_KILL_ZOMBIE_PROCESSES"):
+        is_terminated = self._debug_adapter_robot_target_comm.is_terminated()
+        # i.e.: if the disconnect happens before the RF session sends a terminate
+        # then we need to kill subprocesses (this means the user pressed the
+        # stop button).
+        if is_true_in_env("RFLS_KILL_ZOMBIE_PROCESSES") or not is_terminated:
             from robocorp_ls_core.basic import kill_process_and_subprocesses
 
             if self._popen is not None: